telegram-bot/requirements.txt
# RITMO Telegram Bot Dependencies

# Bot framework
python-telegram-bot[webhooks]==21.0

# LLM: OpenAI GPT
openai>=1.30.0

# HTTP requests 
httpx~=0.27
aiohttp>=3.9.1

# Environment variables
python-dotenv==1.0.0

# Date/time handling
pytz==2023.3

# Data validation
pydantic>=2.7.0

# Logging
structlog==23.2.0

# Database (Supabase client)
supabase==2.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0